+++ src/App.js
@@ -1,6 +1,6 @@
 import { Console } from '@woowacourse/mission-utils';
 import Validation from './modules/Validation.js';
-import Decision from './modules/Decision.js';
+import Print from './modules/Print.js';
 import MESSAGE from './constants/message.js';
 import CONDITION from './constants/condition.js';
 
@@ -32,25 +32,8 @@ class App {
       MESSAGE.input.numberOfTimes,
     );
     this.numberOfTimes = Validation.getNumberOfTimes(answerOfNumber);
-    this.printExecutionResult();
+    this.cars = Print.executeProcess(this.cars, this.numberOfTimes);
     this.printFinalWinner();
-  }
-
-  printExecutionResult() {
-    Console.print(MESSAGE.result.execution);
-    const updatedCars = new Map([...this.cars]);
-    for (let i = 1; i <= this.numberOfTimes; i += 1) {
-      const result = [];
-      updatedCars.forEach((value, key) => {
-        let moveNum = value;
-        if (Decision.moveForward()) {
-          updatedCars.set(key, (moveNum += 1));
-        }
-        result.push(`${key} : ${MESSAGE.result.distance.repeat(moveNum)}\n`);
-      });
-      Console.print(result.join(''));
-    }
-    this.cars = updatedCars;
   }
 
   printFinalWinner() {
